+++ Client/src/components/App.js
@@ -6,7 +6,6 @@ export default function App() {
   return (
     <div>
       <Sidebar/>
-      Hello World..!
     </div>
   )
 }
