+++ Client/src/components/common/Link/index.css
@@ -1,0 +1,10 @@
+.sublink{
+  font-size: 16px;
+  padding: 7px;
+  cursor: pointer;
+  margin-left: 25px;
+}
+
+.sublink:active{
+  border-left: 2px solid #eee;
+}
