+++ Client/src/components/common/Link/index.js
@@ -1,0 +1,5 @@
+import React from 'react';
+import './index.css'
+
+export default function Link({ link }) { return (<li className='sublink'> {link} </li>) };
+
