--- Client/src/components/common/Sidebar/Link/index.js
@@ -1,5 +1,0 @@
-import React from 'react';
-import './index.css'
-
-export default function Link({ link }) { return (<li className='sublink'> {link} </li>) }
-
