--- Client/src/components/common/Sidebar/Linklst/index.css
@@ -1,8 +1,0 @@
-.li-links{
-  padding: 7px;
-}
-
-.link{
-  list-style-type: none;
-  padding-left: 10px;
-}
