+++ Client/src/components/common/Sidebar/index.css
@@ -7,7 +7,7 @@
 .sidebar{
   width:240px;
   background-color:#2980B9;
-  height: 100vh;
+  min-height: 100vh;
   color: #eee;
   font-family:'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
 }
@@ -22,6 +22,6 @@ text-align: center;
   margin: 10px auto;
 }
 .linklst{
-  margin-left: 30px;
+  margin-left: 12px;
   list-style-type: none;
 }
